requirements.txt
celery==4.2.1
Django==2.1
redis==2.10.6
six==1.11.0
telepot==12.7
djangorestframework==3.8.2
